requirements.txt
opencv-python
mediapipe
numpy
SpeechRecognition
pyttsx3
python-dotenv
openwakeword
google-generativeai
